Add fallback value and collection
