+++ library/Zend/Hydrator/HydratorInterface.php
@@ -1,0 +1,32 @@
+<?php
+/**
+ * Zend Framework (http://framework.zend.com/)
+ *
+ * @link      http://github.com/zendframework/zf2 for the canonical source repository
+ * @copyright Copyright (c) 2005-2013 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   http://framework.zend.com/license/new-bsd New BSD License
+ */
+
+namespace Zend\Hydrator;
+
+interface HydratorInterface
+{
+    /**
+     * Extract values from an object
+     *
+     * @param  object $object
+     * @param  bool   $recursive
+     * @return array
+     */
+    public function extract($object, $recursive = true);
+
+    /**
+     * Hydrate $object with the provided $data.
+     *
+     * @param  array  $data
+     * @param  object $object
+     * @param  bool   $recursive
+     * @return object
+     */
+    public function hydrate(array $data, $object, $recursive = true);
+}
